+++ src/components/Profile/Profile.jsx
@@ -4,11 +4,11 @@ import Myposts from "./Myposts/Myposts";
 import ProfileInfo from "./ProfileInfo/ProfileInfo";
 
 
-const Profile = () => {
+const Profile = (props) => {
   return (
     <div>
       <ProfileInfo />
-      <Myposts />
+      <Myposts mypostp={props.mypost}/>
     </div>
   );
 };
